+++ packages/orq-rc/RELEASES.md
@@ -609,3 +609,13 @@ Based on:
 - [python v3.2.0-rc.23] packages/orq-rc
 ### Releases
 - [PyPI v3.2.0-rc.23] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.23 - packages/orq-rc
+
+## 2025-03-03 11:15:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.508.1 (2.536.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.2.0-rc.24] packages/orq-rc
+### Releases
+- [PyPI v3.2.0-rc.24] https://pypi.org/project/orq-ai-sdk/3.2.0-rc.24 - packages/orq-rc